requirements.txt
numpy
scipy
pyopengl
pyopengl-accelerate
pyside2